pyproject.toml
[project]
name = "llm-utils"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "datetime>=5.5",
    "ipykernel>=6.29.5",
    "openai>=1.95.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
